Optimized Arc* a little bit on an event-by-event basis and removed old useless code
